OMAP3:PM: V2 let DSP limit the VDD1 OPP to 1G while running
with 1.3G-65Mhz enabled in the system, we need to take care of the
below case:
- When DSP runs anything beyond 65MHz disable 1.3G in the system.
Basically, hold a max constraint to limit the OPP table. Any request
to 1.3G at this point will be saved and will be given when the DSP load
dies down to 65MHz.

- some logic fix in SRF for get_opp_id(). It wasn't taking care of the case
were the highest OPP has lower freq than the other OPP's.

Change-Id: I6ba4fc54065fb50a957e1c925d0a2e170d5e885c
Signed-off-by: Girish S G <girishsg@ti.com>