OMAP3: PM: Rebased: DSS check vsync before changing ddr rate-v2
This patch fixes FIFO underflow when 32 bpp is enabled.
During DDR rate change, DDR is stalled and if this stall
time is greater than FIFO deleption rate than you will
hit FIFO_UNDERFLOW, as FIFO wouldn't be refilled fast enough
as it has to.
There are couple of option to take care as per TI's recommendation
To brief:
 - Check if fine tunning of FIFO's work
 - or reducing the image size helps.
 - If none of the above option is viable, then the option implimented
here should be considered; this basically tries to perform DDR
rate change/stall in between the frames.

Change-Id: I70617088140162f990c9e4d322d195f6e8c8ecdc
Signed-off-by: Girish S G <girishsg@ti.com>
Acked-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>